+++ test/rendercli/RenderGraphOptionTest.cmake
@@ -2303,6 +2303,11 @@ if(NOT wavefront_metrics_json MATCHES "\"intersectionBackendPlatformGpuDeviceAva
                   "wavefront metrics report did not contain platform GPU device availability"
                   "" "" "${wavefront_metrics_json}" "")
 endif()
+if(NOT wavefront_metrics_json MATCHES "\"intersectionBackendPlatformGpuRenderPathAvailable\"")
+  _rendercli_fail("rendercli wavefront metrics platform GPU render path availability"
+                  "wavefront metrics report did not contain platform GPU render path availability"
+                  "" "" "${wavefront_metrics_json}" "")
+endif()
 if(NOT wavefront_metrics_json MATCHES "\"frontierRay4PacketChunksPerDepth\"")
   _rendercli_fail("rendercli wavefront metrics frontier Ray4 packet chunks"
                   "wavefront metrics report did not contain frontier Ray4 packet chunk counters"
@@ -2635,7 +2640,7 @@ rendercli_run(
   NAME "rendercli reports batched wavefront any-hit visibility metrics"
   OUTPUT_VARIABLE wavefront_batched_visibility_stdout
   STDOUT_MATCHES
-    "wavefront_metrics.*integrator=pathtracer.*execution=depth_major_paths.*intersection_backend_platform=(metal|vulkan).*intersection_backend_gpu_device=(true|false).*any_hit_rays=[1-9][0-9]*.*any_hit_queries=[1-9][0-9]*.*closest_hit_batch_preferred=true.*any_hit_batch_preferred=true.*direct_light_any_hit_batch_chunks=[1-9][0-9]*.*direct_light_any_hit_batch_rays=[1-9][0-9]*.*direct_light_samples=[1-9][0-9]*"
+    "wavefront_metrics.*integrator=pathtracer.*execution=depth_major_paths.*intersection_backend_platform=(metal|vulkan).*intersection_backend_gpu_device=(true|false).*intersection_backend_gpu_render_path=(true|false).*any_hit_rays=[1-9][0-9]*.*any_hit_queries=[1-9][0-9]*.*closest_hit_batch_preferred=true.*any_hit_batch_preferred=true.*direct_light_any_hit_batch_chunks=[1-9][0-9]*.*direct_light_any_hit_batch_rays=[1-9][0-9]*.*direct_light_samples=[1-9][0-9]*"
   COMMAND
     "${RENDERCLI}" --engine wavefront --integrator pathtracer --width 16 --height 16
     --pathtracer_direct_light_samples 3 --wavefront_denoiser none